% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{runs shiny app locally}
\usage{
run_shiny_app()
}
\value{
No return value
}
\description{
This function allows for local execution of the shiny app. Alternatively,
an online version of this app
\href{https://thijsjanzen.shinyapps.io/simRestoreApp/}{can be found here}.
}
