% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_policy.R
\name{simulate_policy}
\alias{simulate_policy}
\title{Simulate the effect of a policy over time.}
\usage{
simulate_policy(
  initial_population_size = 400,
  nest_success_rate = 0.387,
  nesting_risk = 0.2,
  K = 400,
  num_generations = 20,
  pull = 0,
  put = 0,
  starting_freq = 0.5,
  sd_starting_freq = 0.05,
  morgan = 1,
  establishment_burnin = 30,
  num_replicates = 1,
  seed = NULL,
  max_age = 6,
  mean_clutch_size = 6,
  sd_clutch_size = 1,
  smin = 0.5,
  smax = 0.9,
  b = -2,
  p = 0.5,
  sex_ratio_put = 0.5,
  sex_ratio_offspring = 0.5,
  use_simplified_model = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{initial_population_size}{population size at the start}

\item{nest_success_rate}{frequency of nests that yield offspring at the end
of the breeding season (e.g. a fraction of 1 - nest_success_rate of nests
fail). This is a joint effect of breeding females getting killed
(see \code{female_death_rate}) and other sources of failure to complete a
nest. Other sources of failure are calculated from nest_success_rate and
female_death_rate, such that nest failure rate = 1 - nest_success_rate / (1 -
female_death_rate);}

\item{nesting_risk}{Additional death rate of females as a result of
protecting the nest.}

\item{K}{carrying capacity}

\item{num_generations}{number of generations}

\item{pull}{vector of the number of individuals pulled per year}

\item{put}{vector of the number of individuals added per year}

\item{starting_freq}{initial hawaii frequency in the population.}

\item{sd_starting_freq}{variation in initial hawaii frequency.}

\item{morgan}{size of the chromosome in Morgan}

\item{establishment_burnin}{number of generations before establishment}

\item{num_replicates}{number of replicates}

\item{seed}{random number seed, if left open, current time is used.}

\item{max_age}{maximum age a duck can reach.}

\item{mean_clutch_size}{mean number of eggs in a nest}

\item{sd_clutch_size}{standard deviation of number of eggs in nest (assuming
the number of eggs is always 0 or larger).}

\item{smin}{minimum survival rate}

\item{smax}{maximum survival rate}

\item{b}{steepness of the survival rate. Negative values indicate a declining
survival rate with increasing population size, positive values indicate an
increasing survival rate with increasing population size.}

\item{p}{Density at which the survival rate changes most relative. Expressed
in N / K (e.g., for a value of 1.0, the survival rate changes most rapidly
around N = K, for a value of 0.5, the survival rate changes most rapidly
around N = 0.5K, etc).}

\item{sex_ratio_put}{the sex ratio of individuals that are added (if any) to
the population. Sex ratio is expressed as males / (males + females), such
that 0.5 indicates an even sex ratio, 0.9 indicates a male biased sex ratio
and 0.1 indicates a female biased sex ratio.}

\item{sex_ratio_offspring}{sex ratio of newly born offspring. The sex ratio
is expressed as males / (males + females), such
that 0.5 indicates an even sex ratio, 0.9 indicates a male biased sex ratio
and 0.1 indicates a female biased sex ratio.}

\item{use_simplified_model}{use a simplified model of underlying genetics?
This speeds up simulation considerably, and should be preferred when not
interested in high detail genetic changes. For subsequent ADMIXTURE analysis,
use_simplified_model should be set to FALSE. Default is TRUE.}

\item{verbose}{provides verbose output if TRUE.}
}
\value{
tibble
}
\description{
Simulate the effect of a policy over time.
}
