% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{update_vector}
\alias{update_vector}
\title{update input vector}
\usage{
update_vector(v, num_generations)
}
\arguments{
\item{v}{vector or single number}

\item{num_generations}{number of generations}
}
\description{
update input vector
}
\keyword{internal}
