% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_static.R
\name{optimize_static}
\alias{optimize_static}
\title{Optimize putting and/or pulling, where it is assumed that the same amount
is applied per generation.}
\usage{
optimize_static(
  num_generations = 20,
  target_frequency = 0.99,
  K = 400,
  optimize_put = TRUE,
  optimize_pull = FALSE,
  num_replicates = 1,
  use_simplified_model = TRUE,
  verbose = FALSE,
  initial_population_size = 400,
  reproduction_success_rate = 0.387,
  breeding_risk = c(0.2, 0),
  starting_freq = 0.2,
  morgan = 1,
  max_age = 6,
  mean_number_of_offspring = 6,
  sd_number_of_offspring = 1,
  smin = 0.5,
  smax = 0.9,
  b = -2,
  p = 0.5,
  sex_ratio_put = 0.5,
  sex_ratio_offspring = 0.5,
  establishment_burnin = 30
)
}
\arguments{
\item{num_generations}{number of generations}

\item{target_frequency}{frequency to aim for}

\item{K}{carrying capacity}

\item{optimize_put}{When set to 0, FALSE or a negative number, it will not
be optimized. When negative, the absolute value will be taken as a fixed
contribution to each generation (but will not be optimized)}

\item{optimize_pull}{When set to 0, FALSE or a negative number, it will not
be optimized. When negative, the absolute value will be taken as a fixed
contribution to each generation (but will not be optimized)}

\item{num_replicates}{number of replicates}

\item{use_simplified_model}{use a simplified model of underlying genetics?
This speeds up simulation considerably, and should be preferred when not
interested in high detail genetic changes. Default is TRUE.}

\item{verbose}{provides verbose output if TRUE.}

\item{initial_population_size}{population size at the start}

\item{reproduction_success_rate}{frequency of females that yield offspring
at the end of the breeding season (e.g. a fraction of
1 - reproduction_success_rate of females. This is a joint effect of
breeding females getting killed
(see \code{female_death_rate}) and other sources of failure to complete a
clutch. Other sources of failure are calculated from nest_success_rate and
female_death_rate, such that reproduction failure rate =
1 - reproduction_success_rate / (1 - female breeding risk);}

\item{breeding_risk}{Additional death rate of males and females as a result
of breeding (e.g. as a result of protecting the offspring against predators).
Provide as a vector where the first index indicates the risk for females,
the second the risk for males.}

\item{starting_freq}{initial frequency of target ancestry in the population.}

\item{morgan}{vector sizes of chromosomes in Morgan}

\item{max_age}{maximum age an individual can reach.}

\item{mean_number_of_offspring}{mean number of offspring per female}

\item{sd_number_of_offspring}{standard deviation of number of offspring per
female (assuming the number of offspring is always 0 or larger).}

\item{smin}{minimum survival rate}

\item{smax}{maximum survival rate}

\item{b}{steepness of the survival rate. Negative values indicate a declining
survival rate with increasing population size, positive values indicate an
increasing survival rate with increasing population size.}

\item{p}{Density at which the survival rate changes most relative. Expressed
in N / K (e.g., for a value of 1.0, the survival rate changes most rapidly
around N = K, for a value of 0.5, the survival rate changes most rapidly
around N = 0.5K, etc).}

\item{sex_ratio_put}{sex ratio of individuals that are added (if any) to
the population. Sex ratio is expressed as males / (males + females), such
that 0.5 indicates an even sex ratio, 0.9 indicates a male biased sex ratio
and 0.1 indicates a female biased sex ratio.}

\item{sex_ratio_offspring}{sex ratio of newly born offspring. The sex ratio
is expressed as males / (males + females), such
that 0.5 indicates an even sex ratio, 0.9 indicates a male biased sex ratio
and 0.1 indicates a female biased sex ratio.}

\item{establishment_burnin}{number of generations before establishment}
}
\value{
tibble
}
\description{
Optimize putting and/or pulling, where it is assumed that the same amount
is applied per generation.
}
